added TopHat moved hats to accessories removed Sunglasses special Effect
