+++ src/main/java/net/rand/exten/datagen/lang/DeGen.java
@@ -120,6 +120,7 @@ public class DeGen extends FabricLanguageProvider {
         translationBuilder.add(Items_RaEx.GREEN_SODA_CAN, "Grüne Limonade");
         translationBuilder.add(Items_RaEx.RED_SODA_CAN, "Rote Limonade");
         translationBuilder.add(Items_RaEx.SUNGLASSES, "Sonnenbrille");
+        translationBuilder.add(Items_RaEx.TOP_HAT, "Zylinder");
 
         /// Loot Bag
         translationBuilder.add(LootBag_RaEx.EMPTY_LOOT_BAG, "Lehre Beutetasche");
