+++ src/main/java/net/rand/exten/datagen/lang/EnUsGen.java
@@ -120,6 +120,7 @@ public class EnUsGen extends FabricLanguageProvider {
         translationBuilder.add(Items_RaEx.GREEN_SODA_CAN, "Green Soda");
         translationBuilder.add(Items_RaEx.RED_SODA_CAN, "Red Soda");
         translationBuilder.add(Items_RaEx.SUNGLASSES, "Sunglasses");
+        translationBuilder.add(Items_RaEx.TOP_HAT, "Top Hat");
 
         /// Loot Bag
         translationBuilder.add(LootBag_RaEx.EMPTY_LOOT_BAG, "Empty Loot Bag");
