+++ src/main/java/net/rand/exten/item/ItemGroups_RaEx.java
@@ -207,8 +207,9 @@ public class ItemGroups_RaEx {
                         entries.add(Blocks_RaEx.EXPLOSIVE_CAKE_BLOCK);
                         entries.add(Blocks_RaEx.BARBED_WIRE);
 
-                        /// Hats
+                        /// Head Wearable
                         entries.add(Items_RaEx.SUNGLASSES);
+                        entries.add(Items_RaEx.TOP_HAT);
 
                         /// Funny Items and Blocks
                         entries.add(Items_RaEx.CHEESE_OF_FLIGHT);
@@ -248,6 +249,4 @@ public class ItemGroups_RaEx {
     public static void registerItemGroups() {
         RandomExtensions.LOGGER.info("Register Item Groups for" + RandomExtensions.MOD_ID);
     }
-
-
 }
