+++ src/main/java/net/rand/exten/item/custom/wearable/accessories/head/TopHat.java
@@ -1,0 +1,17 @@
+package net.rand.exten.item.custom.wearable.accessories.head;
+
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.world.World;
+import net.rand.exten.item.custom.wearable.Hats;
+
+public class TopHat extends Hats {
+    public TopHat(Settings settings) {
+        super(settings);
+    }
+
+    @Override
+    public void SpecialAbility(PlayerEntity user, ItemStack stack, World world) {
+
+    }
+}
