+++ src/main/resources/assets/randexten/models/item/top_hat.json
@@ -1,0 +1,86 @@
+{
+	"credit": "Made with Blockbench",
+	"texture_size": [64, 64],
+	"textures": {
+		"0": "randexten:item/top_hat",
+		"1": "item/feather",
+		"particle": "randexten:item/top_hat"
+	},
+	"elements": [
+		{
+			"from": [1, 0, 1],
+			"to": [15, 1, 15],
+			"faces": {
+				"north": {"uv": [0, 14, 7, 14.5], "texture": "#0"},
+				"east": {"uv": [0, 14, 7, 14.5], "texture": "#0"},
+				"south": {"uv": [0, 14, 7, 14.5], "texture": "#0"},
+				"west": {"uv": [0, 14, 7, 14.5], "texture": "#0"},
+				"up": {"uv": [7, 7, 0, 0], "texture": "#0"},
+				"down": {"uv": [7, 7, 0, 14], "texture": "#0"}
+			}
+		},
+		{
+			"from": [3, 1, 3],
+			"to": [13, 10, 13],
+			"faces": {
+				"north": {"uv": [7, 0, 12, 4.5], "texture": "#0"},
+				"east": {"uv": [7, 4.5, 12, 9], "texture": "#0"},
+				"south": {"uv": [7, 9, 12, 13.5], "texture": "#0"},
+				"west": {"uv": [12, 4.5, 7, 9], "texture": "#0"},
+				"up": {"uv": [6, 6, 1, 1], "texture": "#0"}
+			}
+		},
+		{
+			"from": [2.95, 5, 8],
+			"to": [2.95, 13, 16],
+			"faces": {
+				"north": {"uv": [0, 0, 0, 4], "texture": "#0"},
+				"east": {"uv": [16, 0, 0, 16], "texture": "#1"},
+				"south": {"uv": [0, 0, 0, 4], "texture": "#0"},
+				"west": {"uv": [0, 0, 16, 16], "texture": "#1"},
+				"up": {"uv": [0, 4, 0, 0], "texture": "#0"},
+				"down": {"uv": [0, 0, 0, 4], "texture": "#0"}
+			}
+		}
+	],
+	"display": {
+		"thirdperson_righthand": {
+			"rotation": [-35, 0, 0],
+			"translation": [0, 2, -1.5],
+			"scale": [0.55, 0.55, 0.55]
+		},
+		"thirdperson_lefthand": {
+			"rotation": [-35, 0, 0],
+			"translation": [0, 2, -1.5],
+			"scale": [0.55, 0.55, 0.55]
+		},
+		"firstperson_righthand": {
+			"rotation": [-35, 0, 0],
+			"translation": [0, 2, -1.5],
+			"scale": [0.55, 0.55, 0.55]
+		},
+		"firstperson_lefthand": {
+			"rotation": [-35, 0, 0],
+			"translation": [0, 2, -1.5],
+			"scale": [0.55, 0.55, 0.55]
+		},
+		"ground": {
+			"translation": [0, 2, 0],
+			"scale": [0.5, 0.5, 0.5]
+		},
+		"gui": {
+			"rotation": [10, 25, 0],
+			"translation": [0, 4, 0],
+			"scale": [0.85, 0.85, 0.85]
+		},
+		"head": {
+			"rotation": [6.25, 0, 0],
+			"translation": [0, 13.75, 1.25],
+			"scale": [1.45, 1.45, 1.45]
+		},
+		"fixed": {
+			"rotation": [0, -90, 0],
+			"translation": [0, 4.75, 0]
+		}
+	}
+}
